+++ src/components/articles/ArticleDelete.vue
@@ -38,7 +38,7 @@
 
 <script>
 import Lang from "@/settings/lang";
-import { apiRequest } from "@/api/Api";
+import { apiRequest } from "@/api/api";
 import { API_ROUTES } from "@/settings/api";
 
 export default {
