+++ src/components/articles/ArticleList.vue
@@ -41,7 +41,7 @@
 import ArticleCard from "@/components/articles/ArticleCard";
 import ArticleDelete from "@/components/articles/ArticleDelete";
 import Lang from "@/settings/lang";
-import { apiRequest } from "@/api/Api";
+import { apiRequest } from "@/api/api";
 import { API_ROUTES } from "@/settings/api";
 
 export default {
